# Optimise `GetBoundCopy` to not rely on reflection to create bindable instances