--- a/osu.Framework.Tests/Bindables/BindableBindingTest.cs
+++ b/osu.Framework.Tests/Bindables/BindableBindingTest.cs
@@ -599,6 +599,8 @@ namespace osu.Framework.Tests.Bindables
                 base.UnbindFrom(them);
                 IsBound = false;
             }
+
+            protected internal override Bindable<int> CreateInstance() => new TestCustomBindable();
         }
     }
 }