--- a/osu.Framework.Tests/Bindables/RangeConstrainedBindableTest.cs
+++ b/osu.Framework.Tests/Bindables/RangeConstrainedBindableTest.cs
@@ -107,6 +107,8 @@ namespace osu.Framework.Tests.Bindables
             }
 
             protected override int DefaultMaxValue => 1;
+
+            protected internal override Bindable<int> CreateInstance() => new BindableNumberWithDefaultMaxValue();
         }
 
         private class BindableNumberWithDefaultMinValue : BindableInt
@@ -117,6 +119,8 @@ namespace osu.Framework.Tests.Bindables
             }
 
             protected override int DefaultMinValue => 3;
+
+            protected internal override Bindable<int> CreateInstance() => new BindableNumberWithDefaultMinValue();
         }
     }
 }