--- a/osu.Framework/Bindables/BindableBool.cs
+++ b/osu.Framework/Bindables/BindableBool.cs
@@ -24,6 +24,6 @@ namespace osu.Framework.Bindables
 
         public void Toggle() => Value = !Value;
 
-        protected internal override Bindable<bool> CreateInstance() => new BindableBool();
+        protected override Bindable<bool> CreateInstance() => new BindableBool();
     }
 }