--- a/osu.Framework/Bindables/BindableDictionary.cs
+++ b/osu.Framework/Bindables/BindableDictionary.cs
@@ -497,21 +497,12 @@ namespace osu.Framework.Bindables
 
         private void removeWeakReference(WeakReference<BindableDictionary<TKey, TValue>> weakReference) => bindings?.Remove(weakReference);
 
-        IBindable IBindable.GetBoundCopy() => GetBoundCopy();
+        IBindable IBindable.CreateInstance() => new BindableDictionary<TKey, TValue>();
 
-        IBindableDictionary<TKey, TValue> IBindableDictionary<TKey, TValue>.GetBoundCopy()
-            => GetBoundCopy();
+        IBindableDictionary<TKey, TValue> IBindableDictionary<TKey, TValue>.GetBoundCopy() => GetBoundCopy();
 
-        /// <summary>
-        /// Create a new instance of <see cref="BindableDictionary{TKey, TValue}"/> and binds it to this instance.
-        /// </summary>
-        /// <returns>The created instance.</returns>
-        public BindableDictionary<TKey, TValue> GetBoundCopy()
-        {
-            var copy = new BindableDictionary<TKey, TValue>();
-            copy.BindTo(this);
-            return copy;
-        }
+        /// <inheritdoc cref="IBindable.GetBoundCopy"/>
+        public BindableDictionary<TKey, TValue> GetBoundCopy() => (BindableDictionary<TKey, TValue>)((IBindable)this).GetBoundCopy();
 
         #endregion IBindableCollection