--- a/osu.Framework/Bindables/BindableDouble.cs
+++ b/osu.Framework/Bindables/BindableDouble.cs
@@ -13,5 +13,7 @@ namespace osu.Framework.Bindables
         }
 
         public override string ToString() => Value.ToString("0.0###", NumberFormatInfo.InvariantInfo);
+
+        protected internal override Bindable<double> CreateInstance() => new BindableDouble();
     }
 }