--- a/osu.Framework/Bindables/BindableInt.cs
+++ b/osu.Framework/Bindables/BindableInt.cs
@@ -13,5 +13,7 @@ namespace osu.Framework.Bindables
         }
 
         public override string ToString() => Value.ToString(NumberFormatInfo.InvariantInfo);
+
+        protected internal override Bindable<int> CreateInstance() => new BindableInt();
     }
 }