--- a/osu.Framework/Bindables/BindableMarginPadding.cs
+++ b/osu.Framework/Bindables/BindableMarginPadding.cs
@@ -62,5 +62,7 @@ namespace osu.Framework.Bindables
                    min.Bottom <= max.Bottom &&
                    min.Right <= max.Right;
         }
+
+        protected internal override Bindable<MarginPadding> CreateInstance() => new BindableMarginPadding();
     }
 }