--- a/osu.Framework/Bindables/BindableNumber.cs
+++ b/osu.Framework/Bindables/BindableNumber.cs
@@ -317,6 +317,8 @@ namespace osu.Framework.Bindables
             }
         }
 
+        protected internal override Bindable<T> CreateInstance() => new BindableNumber<T>();
+
         protected sealed override T ClampValue(T value, T minValue, T maxValue) => max(minValue, min(maxValue, value));
 
         protected sealed override bool IsValidRange(T min, T max) => min.CompareTo(max) <= 0;