--- a/osu.Framework/Bindables/BindableNumberWithCurrent.cs
+++ b/osu.Framework/Bindables/BindableNumberWithCurrent.cs
@@ -31,5 +31,7 @@ namespace osu.Framework.Bindables
             : base(defaultValue)
         {
         }
+
+        protected internal override Bindable<T> CreateInstance() => new BindableNumberWithCurrent<T>();
     }
 }