--- a/osu.Framework/Bindables/BindableSafeArea.cs
+++ b/osu.Framework/Bindables/BindableSafeArea.cs
@@ -16,6 +16,6 @@ namespace osu.Framework.Bindables
         {
         }
 
-        protected internal override Bindable<MarginPadding> CreateInstance() => new BindableSafeArea();
+        protected override Bindable<MarginPadding> CreateInstance() => new BindableSafeArea();
     }
 }