--- a/osu.Framework/Bindables/BindableSize.cs
+++ b/osu.Framework/Bindables/BindableSize.cs
@@ -40,6 +40,8 @@ namespace osu.Framework.Bindables
             }
         }
 
+        protected internal override Bindable<Size> CreateInstance() => new BindableSize();
+
         protected sealed override Size ClampValue(Size value, Size minValue, Size maxValue)
         {
             return new Size