# Breaking changes

## All custom implementations of `IBindable` must implement `CreateInstance()`

Bindables previously used `Activator.CreateInstance()` to implement the `GetBoundCopy()` call. In profiling this has turned out to be a bottleneck in some scenarios, so in order to reduce the associated runtime overhead to about half, all implementors of `IBindable` now must implement `CreateInstance()`.

This also applies to all inheritors of framework-provided bindable types, who must instead override `CreateInstance()` from their base types.

For a given leaf bindable type in the inheritance hierarchy, the method should return a new constructed instance of the leaf type. The recommended pattern to use is as follows:

```csharp
public class BaseBindable : IBindable
{
    IBindable IBindable.CreateInstance() => CreateInstance();
    protected virtual BaseBindable CreateInstance() => new BaseBindable();

    IBindable IBindable.GetBoundCopy() => GetBoundCopy();
    public BaseBindable GetBoundCopy()
        => IBindable.GetBoundCopyImplementation(this); // automatically calls CreateInstance()
}

public sealed class CustomBindable : BaseBindable
{
    protected override BaseBindable CreateInstance() => new CustomBindable();
}
```

---

Closes #4685 

|          Method |     Mean |     Error |    StdDev | Ratio | RatioSD |
|---------------- |---------:|----------:|----------:|------:|--------:|
| GetBoundCopyOld | 2.352 us | 0.0425 us | 0.0840 us |  1.00 |    0.00 |
|    GetBoundCopy | 1.071 us | 0.0212 us | 0.0343 us |  0.46 |    0.02 |

While this avoids reflection-based metods of creating a bindable instance, the new protected `CreateInstance` method has to check that the returned instance matches the same type as the original bindable which is to be bound against, and throw if it doesn't match.

Doing that causes JIT to disable inlining on the method (see https://github.com/ppy/osu-framework/issues/3470). I've defined a `ThrowHelper` class for that purpose, as otherwise `GetBoundCopy` would've been less performant.

---

The fact that all inheritors of `Bindable`s need to override `CreateInstance` is a bit disturbing, but since the chances of making a new bindable class is rare, I believe it's worth the optimisation.

I've left `LocalisedBindableString` from implementing `CreateInstance` on it as it doesn't seem possible to do so, and it's already broken before so I don't see any way of supporting that currently. Leaving it for later when deemed required.