@peppy 2021-08-15
Why does the `throw` need to be there? Is this intended to protect third parties that may be implementing their own bindable incorrectly? Not sure we need that.