@frenzibyte 2021-08-15
> Why does the `throw` need to be there? Is this intended to protect third parties that may be implementing their own bindable incorrectly? Not sure we need that.

Added that for two reasons:
 1. Would act same as what happens when invoking `Activator.CreateInstance` on a bindable which doesn't implement a matching constructor.
 2. Avoids the chances of forgetting to override `CreateInstance`, which can happen silently and result in unexpected behaviour. I can imagine that potentially happening osu!-side, with bindables that override the `Value` setter for example.

That said, I'm wondering if there's a reason you're against having it?