@frenzibyte 2021-08-15
> In addition, I'm not sure why you've listed four types of invocation in the benchmark when there should only be two compared - what we were using and what the new code is providing. It's unclear immediately which two we should be looking at (guessing the first two, where the... second is the old?), especially since the `baseline` is set incorrectly to the new one, rather than the old.
> 
> Here's a better way (in my opinion) to create benchmark comparisons for PR descriptions:
> 
> * Copy the pre-changes version of `Bindable` to `BindableOld`
> * Create two benchmark methods with the same content, except that one instantiates a `Bindable` and the other `BindableOld`. Set the old one as the baseline

I thought `Baseline` is supposed to be set on the pure way of creating a bindable, then the other benchmarks would scale based on it to show the ratio between the pure way and the other methods. But perhaps not.

Your way sounds good, I can't exactly copy the pre-changed version of `Bindable`, but I can at least make a bindable that overrides to create the instance via `Activator` to reflect the overhead.