@frenzibyte 2021-08-15
Benchmarks updated, shows a ratio of 0.50 on the new way.