@peppy 2021-08-15
I guess the throwing is required due to the `virtual` implementation in `Bindable`, seems fine to leave it.

The new benchmarks make me question whether this was required in the first place. At 2.3us this method likely does not need this change.