@smoogipoo 2021-08-17
I've refactored this to remove the back-and-forth casting.

* `IBindable` exposes a `GetBoundCopy()` method.
* Custom interfaces like `IBindable<T>` can expose their own typed variant of `GetBoundCopy()`
* The final implementation (e.g. `Bindable<T>`) implements `GetBoundCopy()` via `IBindable.GetBoundCopyImplementation(this)`.
* `IBindable.GetBoundCopyImplementation()` is only available within `IBindable` subclasses.
* `IBindable.CreateInstance()` is now only `protected` instead of `protected internal`.

|          Method |       Mean |   Error |  StdDev | Ratio |
|---------------- |-----------:|--------:|--------:|------:|
| GetBoundCopyOld | 1,729.3 ns | 2.92 ns | 2.59 ns |  1.00 |
|    GetBoundCopy |   797.9 ns | 7.01 ns | 6.22 ns |  0.46 |

Relative performance hasn't changed from the OP.