@bdach 2021-08-17
Code seems good to me. I've added a breaking change note to the PR description.

@frenzibyte before I go merging this, do you have the corresponding game-side changes ready? because this *will* break custom bindable types and the game has a few of those.